Big code reorg. Created stack.py and utils.py
